requirements.txt
requests>=2.31.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
replicate>=0.25.0
edge-tts>=6.1.0
pygame>=2.5.0
